# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/magic-movers":{"post":{"operationId":"MagicMoversController_addMover","summary":"Add a new Magic Mover","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMagicMoverDto"}}}},"responses":{"201":{"description":"The magic mover has been successfully created."}},"tags":["magic-movers"]}},"/magic-movers/{moverId}/load":{"post":{"operationId":"MagicMoversController_loadMover","summary":"Load items to a Magic Mover","parameters":[{"name":"moverId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoadItemsDto"}}}},"responses":{"200":{"description":"Items loaded successfully."}},"tags":["magic-movers"]}},"/magic-movers/{moverId}/start-mission":{"post":{"operationId":"MagicMoversController_startMission","summary":"Start a mission for a Magic Mover","parameters":[{"name":"moverId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Mission started successfully."}},"tags":["magic-movers"]}},"/magic-movers/{moverId}/end-mission":{"post":{"operationId":"MagicMoversController_endMission","summary":"End a mission for a Magic Mover","parameters":[{"name":"moverId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Mission ended successfully."}},"tags":["magic-movers"]}},"/magic-movers/leaderboard":{"get":{"operationId":"MagicMoversController_getLeaderboard","summary":"Get leaderboard of Magic Movers","parameters":[],"responses":{"200":{"description":"Leaderboard retrieved successfully."}},"tags":["magic-movers"]}}},"info":{"title":"Magic-Mover","description":"The Magic-Mover API description","version":"1.0","contact":{}},"tags":[{"name":"transport","description":""}],"servers":[],"components":{"schemas":{"CreateMagicMoverDto":{"type":"object","properties":{"weightLimit":{"type":"number","example":100,"description":"The maximum weight limit the mover can carry"},"energy":{"type":"number","example":50,"description":"The total magic power of the mover"}},"required":["weightLimit","energy"]},"LoadItemsDto":{"type":"object","properties":{"itemIds":{"example":[1,2],"description":"List of item IDs to load","type":"array","items":{"type":"string"}}},"required":["itemIds"]}}}}